requirements.txt
llama-cpp-python
orjson
pydantic==2.*
rich
pytest
fastapi